add chat

+++ lib/app/modules/detail_appointment/controllers/detail_appointment_controller.dart
@@ -1,3 +1,4 @@
+import 'package:client_mohamoon/app/models/appointment_model.dart';
 import 'package:flutter_easyloading/flutter_easyloading.dart';
 
 //import 'package:flutter_stripe/flutter_stripe.dart';
@@ -19,10 +20,10 @@ class DetailAppointmentController extends GetxController {
   final UserService userService = Get.find();
   final TimeSlotService timeSlotService = Get.find();
   List<AppointmentDetailModel> appointmentDetail = List.empty();
-  TimeSlot selectedTimeSlot = Get.arguments[0];
+  dynamic selectedTimeSlot = Get.arguments[0];
   Lawyer lawyer = Get.arguments[1];
   int duration = Get.arguments[2];
-  List<TimeSlot> timeSlots = Get.arguments[3];
+  dynamic timeSlots = Get.arguments[3];
   PaymentService paymentService = Get.find();
   NotificationService notificationService = Get.find<NotificationService>();
   late String clientSecret;
@@ -34,24 +35,34 @@ class DetailAppointmentController extends GetxController {
     userService.getUsername().then((value) {
       username.value = value;
     });
-    price = selectedTimeSlot.price!;
-    price = duration / 15 * price;
+    price = (lawyer.lawyerPrice)!.toDouble();
+    price = price / 4;
   }
 
   @override
   void onClose() {}
 
-  AppointmentDetailModel buildAppointmentDetail() {
-    var formatter = DateFormat('yyyy-MM-dd hh:mm');
-    var time = formatter.format(selectedTimeSlot.timeSlot!);
+  AppointmentDetailModel buildOrderDetail() {
+    if (selectedTimeSlot != null) {
+      var formatter = DateFormat('yyyy-MM-dd hh:mm');
+      var time = formatter.format(selectedTimeSlot.timeSlot!);
 
-    var appointmentDetail = AppointmentDetailModel(
-        itemId: selectedTimeSlot.timeSlotId!,
-        itemName: 'Consultation with ${lawyer.lawyerName!}',
-        time: time,
-        duration: '${duration} minute',
-        price: currencySign + price.toString());
-    return appointmentDetail;
+      var appointmentDetail = AppointmentDetailModel(
+          itemId: selectedTimeSlot.timeSlotId!,
+          itemName: 'Consultation with ${lawyer.lawyerName!}',
+          time: time,
+          duration: '$duration minute',
+          price: currencySign + price.toString());
+      return appointmentDetail;
+    } else {
+      var appointmentDetail = AppointmentDetailModel(
+          itemId: "0",
+          itemName: 'Consultation with ${lawyer.lawyerName!}',
+          time: DateTime.now().toString(),
+          duration: '${duration} minute',
+          price: currencySign + price.toString());
+      return appointmentDetail;
+    }
   }
 
   void makePayment() async {
@@ -59,28 +70,50 @@ class DetailAppointmentController extends GetxController {
     try {
       String? userId = userService.currentUser!.uid;
       double? balance = await userService.getUserBalance(userId);
-     /* print(balance);
+      /* print(balance);
       balance = 0;
       print(balance);*/
       if (balance! >= price!) {
-        await paymentService.makePayment(selectedTimeSlot.timeSlotId!,
-            userService.getUserId(), price!, duration!);
-        print('#####');
-        print(timeSlots.length);
-        for (int i = 0; i < timeSlots.length; i++) {
-          print(timeSlots[i].timeSlot);
-          await timeSlotService.deleteTimeSlot(timeSlots[i]);
+        if (selectedTimeSlot != null) {
+          await paymentService.makePayment(selectedTimeSlot.timeSlotId!,
+              userService.getUserId(), price, duration);
+          for (int i = 0; i < timeSlots.length; i++) {
+            print(timeSlots[i].timeSlot);
+            await timeSlotService.deleteTimeSlot(timeSlots[i]);
+          }
+          EasyLoading.dismiss();
+          Get.offNamed(
+            '/payment-success',
+            arguments: price,
+          );
+          notificationService
+              .setNotificationAppointment(selectedTimeSlot.timeSlot!);
+        } else {
+          await paymentService.makePayment(
+            "0",
+            userService.getUserId(),
+            price,
+            duration,
+          );
+
+          EasyLoading.dismiss();
+          print("SSSS");
+          print(lawyer.lawyerId);
+          Get.offNamed(
+            '/appointment-detail',
+            arguments: [
+              null,
+              lawyer,
+            ],
+          );
         }
-        EasyLoading.dismiss();
-        Get.offNamed('/payment-success', arguments: [selectedTimeSlot, price]);
-        notificationService
-            .setNotificationAppointment(selectedTimeSlot.timeSlot!);
       } else {
         Fluttertoast.showToast(msg: 'please charge your balance!');
         EasyLoading.dismiss();
         Get.to(() => BalanceView());
       }
     } catch (err) {
+      EasyLoading.dismiss();
       Fluttertoast.showToast(msg: err.toString());
       return null;
     }
